+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod errors;
 
 pub mod meshgen;
 pub mod lod;
+pub mod presets;
 
 use bevy::{ecs::{lifecycle::HookContext, world::DeferredWorld}, prelude::*};
 use fastrand::Rng;